refactor: move stories normalize and denormalize functions from models to reducers
