+++ app/assets/javascripts/models/beta/story.js
@@ -469,48 +469,3 @@ export const donePoints = (stories) =>
 
 export const remainingPoints = (stories) =>
   totalPoints(stories) - donePoints(stories);
-
-export const normalizeStories = (stories) => {
-  return stories.reduce(
-    (acc, story) => {
-      const storyId = story.id;
-
-      acc.stories.byId[storyId] = { ...story };
-      acc.stories.allIds.push(storyId);
-
-      return acc;
-    },
-    {
-      stories: {
-        byId: {},
-        allIds: [],
-      },
-    }
-  );
-};
-
-export const denormalizeStories = (stories) => {
-  const normalizedStories = stories.stories;
-
-  if (!normalizedStories || normalizedStories.allIds.length === 0) {
-    return [];
-  }
-
-  const denormalizedStories = normalizedStories.allIds.map((storyId) => {
-    return normalizedStories.byId[storyId];
-  });
-
-  return denormalizedStories;
-};
-
-export const denormalizeState = (state) => ({
-  epic: denormalizeStories(state[storyScopes.EPIC]),
-  all: denormalizeStories(state[storyScopes.ALL]),
-  search: denormalizeStories(state[storyScopes.SEARCH]),
-});
-
-export const normalizeState = (state) => ({
-  epic: normalizeStories(state[storyScopes.EPIC]),
-  all: normalizeStories(state[storyScopes.ALL]),
-  search: normalizeStories(state[storyScopes.SEARCH]),
-});
