+++ spec/javascripts/models/beta/story_spec.js
@@ -1476,133 +1476,4 @@ describe("Story model", function () {
       });
     });
   });
-  describe("normalizeStories", () => {
-    it("normalize an array of stories by their IDs", () => {
-      const symbolKey = createTemporaryId();
-      const stories = [{ id: 1 }, { id: 2 }, { id: symbolKey }];
-      const normalizedStories = normalizeStories(stories);
-
-      expect(normalizedStories).toEqual({
-        stories: {
-          byId: {
-            1: { id: 1 },
-            2: { id: 2 },
-            [symbolKey]: { id: symbolKey },
-          },
-          allIds: [1, 2, symbolKey],
-        },
-      });
-    });
-  });
-  describe("denormalizeStories", () => {
-    it("denormalize a set of stories, including a Symbol as a key", () => {
-      const symbolKey = createTemporaryId();
-      const normalizedStories = {
-        stories: {
-          byId: {
-            1: { id: 1 },
-            2: { id: 2 },
-            [symbolKey]: { id: symbolKey },
-          },
-          allIds: [1, 2, symbolKey],
-        },
-      };
-      const denormalizedStories = denormalizeStories(normalizedStories);
-
-      expect(denormalizedStories).toEqual([
-        { id: 1 },
-        { id: 2 },
-        { id: symbolKey },
-      ]);
-    });
-    it("handle an empty object", () => {
-      const normalizedStories = {
-        stories: {
-          byId: {},
-          allIds: [],
-        },
-      };
-      const denormalizedStories = denormalizeStories(normalizedStories);
-
-      expect(denormalizedStories).toEqual([]);
-    });
-  });
-  describe("normalizeState", () => {
-    it("normalize the state object", () => {
-      const symbolKey = Symbol("new");
-      const state = {
-        epic: [{ id: 2 }],
-        all: [{ id: 1 }],
-        search: [{ id: symbolKey }],
-      };
-      const normalizedState = normalizeState(state);
-      const expectedNormalizedState = {
-        epic: {
-          stories: {
-            byId: {
-              2: { id: 2 },
-            },
-            allIds: [2],
-          },
-        },
-        all: {
-          stories: {
-            byId: {
-              1: { id: 1 },
-            },
-            allIds: [1],
-          },
-        },
-        search: {
-          stories: {
-            byId: {
-              [symbolKey]: { id: symbolKey },
-            },
-            allIds: [symbolKey],
-          },
-        },
-      };
-
-      expect(normalizedState).toEqual(expectedNormalizedState);
-    });
-  });
-  describe("denormalizeState", () => {
-    it("denormalize the state object", () => {
-      const symbolKey = Symbol("new");
-      const normalizedState = {
-        epic: {
-          stories: {
-            byId: {
-              2: { id: 2 },
-            },
-            allIds: [2],
-          },
-        },
-        all: {
-          stories: {
-            byId: {
-              1: { id: 1 },
-            },
-            allIds: [1],
-          },
-        },
-        search: {
-          stories: {
-            byId: {
-              [symbolKey]: { id: symbolKey },
-            },
-            allIds: [symbolKey],
-          },
-        },
-      };
-      const denormalizedState = denormalizeState(normalizedState);
-      const expectedDenormalizedState = {
-        epic: [{ id: 2 }],
-        all: [{ id: 1 }],
-        search: [{ id: symbolKey }],
-      };
-
-      expect(denormalizedState).toEqual(expectedDenormalizedState);
-    });
-  });
 });
